ingestor/requirements.txt
redis>=5.0.0
websocket-client>=1.6.0
